Replace OHLC

+++ src/test/java/market/Market.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 
 public interface Market extends SimulationAware {
 
-    long getPrice(Instrument instrument);
+    double getPrice(Instrument instrument);
 
     boolean isAvailable();
